renamed files to standard naming
